divide stories into separate files
